requirements.txt
Django==5.2.1
python-dotenv==1.0.1
requests==2.31.0
google-cloud-aiplatform